+++ src/project/Makefile
@@ -18,7 +18,7 @@ list-service:
 list-project:
 	grpcurl \
 		-plaintext \
-		-d '{"user_id":1001, "project_id":1001, "name":"project-a"}' \
+		-d '{"user_id":1002, "project_id":1001, "name":"project-a"}' \
 		localhost:8003 core.project.ProjectService.ListProject
 
 .PHONY: create-project
@@ -41,3 +41,24 @@ delete-project:
 		-plaintext \
 		-d '{"project_id":1005}' \
 		localhost:8003 core.project.ProjectService.DeleteProject
+
+.PHONY: tag-project
+tag-project:
+	grpcurl \
+		-plaintext \
+		-d '{"project_id":1001, "tag":"key:value"}' \
+		localhost:8003 core.project.ProjectService.TagProject
+
+.PHONY: tag-project2
+tag-project2:
+	grpcurl \
+		-plaintext \
+		-d '{"project_id":1001, "tag":"key:value2"}' \
+		localhost:8003 core.project.ProjectService.TagProject
+
+.PHONY: untag-project
+untag-project:
+	grpcurl \
+		-plaintext \
+		-d '{"project_id":1001, "tag":"key:value"}' \
+		localhost:8003 core.project.ProjectService.UntagProject
