+++ src/project/go.mod
@@ -3,20 +3,21 @@ module github.com/CyberAgent/src/project
 go 1.16
 
 require (
-	github.com/CyberAgent/mimosa-core/pkg/model v0.0.0-20210304051836-3872fb985505
-	github.com/CyberAgent/mimosa-core/proto/iam v0.0.0-20210304051836-3872fb985505
-	github.com/CyberAgent/mimosa-core/proto/project v0.0.0-20210304051836-3872fb985505
+	github.com/CyberAgent/mimosa-core/pkg/model v0.0.0-20210308084022-523df78e8f76
+	github.com/CyberAgent/mimosa-core/proto/iam v0.0.0-20210308084022-523df78e8f76
+	github.com/CyberAgent/mimosa-core/proto/project v0.0.0-20210308084022-523df78e8f76
+	github.com/envoyproxy/protoc-gen-validate v0.4.1 // indirect
 	github.com/go-ozzo/ozzo-validation/v4 v4.3.0 // indirect
 	github.com/go-sql-driver/mysql v1.5.0
 	github.com/golang/protobuf v1.4.3
 	github.com/jinzhu/gorm v1.9.15
 	github.com/kelseyhightower/envconfig v1.4.0
 	github.com/sirupsen/logrus v1.6.0
-	github.com/stretchr/testify v1.5.1
+	github.com/stretchr/testify v1.6.1
 	github.com/vikyd/zero v0.0.0-20190921142904-0f738d0bc858
 	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9 // indirect
 	golang.org/x/net v0.0.0-20210226172049-e18ecbb05110 // indirect
-	golang.org/x/sys v0.0.0-20210303074136-134d130e1a04 // indirect
+	golang.org/x/sys v0.0.0-20210305230114-8fe3ee5dd75b // indirect
 	golang.org/x/text v0.3.5 // indirect
 	google.golang.org/genproto v0.0.0-20210303154014-9728d6b83eeb // indirect
 	google.golang.org/grpc v1.36.0
