+++ src/project/service_tag.go
@@ -1,0 +1,39 @@
+package main
+
+import (
+	"context"
+
+	"github.com/CyberAgent/mimosa-core/pkg/model"
+	"github.com/CyberAgent/mimosa-core/proto/project"
+	"github.com/golang/protobuf/ptypes/empty"
+)
+
+func convertProjectTag(p *model.ProjectTag) *project.ProjectTag {
+	return &project.ProjectTag{
+		ProjectId: p.ProjectID,
+		Tag:       p.Tag,
+		CreatedAt: p.CreatedAt.Unix(),
+		UpdatedAt: p.UpdatedAt.Unix(),
+	}
+}
+
+func (p *projectService) TagProject(ctx context.Context, req *project.TagProjectRequest) (*project.TagProjectResponse, error) {
+	if err := req.Validate(); err != nil {
+		return nil, err
+	}
+	tag, err := p.repository.TagProject(req.ProjectId, req.Tag)
+	if err != nil {
+		return nil, err
+	}
+	return &project.TagProjectResponse{Project: convertProjectTag(tag)}, nil
+}
+
+func (p *projectService) UntagProject(ctx context.Context, req *project.UntagProjectRequest) (*empty.Empty, error) {
+	if err := req.Validate(); err != nil {
+		return nil, err
+	}
+	if err := p.repository.UntagProject(req.ProjectId, req.Tag); err != nil {
+		return nil, err
+	}
+	return &empty.Empty{}, nil
+}
